requirements.txt
accelerate==1.3.0
diffusers==0.33.1
fashionpedia==1.1
huggingface-hub==0.28.1
matplotlib==3.10.0
notebook==7.3.2
numpy==2.1.2
opencv-python==4.11.0.86
pillow==11.0.0
pycocotools==2.0.8
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.48.3
